Resolve conflicts caused by deletion of cosmetic.ts
